fix: address PR #56 Copilot review
Two follow-ups from the post-review scan on the earlier mypy cleanup:

1. Package the type stub.  '_fast.pyi' and 'py.typed' were not picked
   up by the wheel build because setuptools.packages.find only
   grabs .py files by default, and MANIFEST.in listed neither.  Add
   both to MANIFEST.in (sdist) and declare them in
   tool.setuptools.package-data (wheel).  Verified the wheel now
   ships snapvec/_fast.pyi and snapvec/py.typed, so downstream
   mypy/pyright users get the types we added locally.

2. Defensive scalar promotion in probe_scores_l2_monotone.  Modern
   numpy (NEP 50) keeps '2.0 * float32_array' as float32, but older
   numpy promoted it to float64, which would have made the cast()
   annotation a lie at runtime.  Use np.float32(2.0) so the
   expression type-stays-put on both old and new numpy.  No behaviour
   change on numpy 2.x.

+++ snapvec/_kmeans.py
@@ -108,9 +108,13 @@ def probe_scores_l2_monotone(
     probe time gives slightly lower recall, especially with uneven
     cluster sizes.
     """
+    # np.float32(2.0) guards against NEP 50-era numpy upcasting a
+    # Python '2.0' scalar to float64 here; on numpy >= 2.0 this is a
+    # no-op, on older numpy it keeps the return dtype matching the
+    # annotation.
     return cast(
         "NDArray[np.float32]",
-        2.0 * (coarse @ q) - (coarse ** 2).sum(1),
+        np.float32(2.0) * (coarse @ q) - (coarse ** 2).sum(1),
     )
